+++ CHANGELOG.md
@@ -1,5 +1,10 @@
 # Changelog
 
+## [0.0.0-20241123] - 2024-11-23
+- Lexer and regex tests.
+- `match` method in `CompiledRegex` match regular expression to string exactly.
+- Range repetition in regular expression (`re{m,n}`).
+
 ## [0.0.0-20241122] - 2024-11-22
 - Working DFA implementation for regular expression matching.
 - Base incremental lexer working and tested.